requirements.txt
mkdocs==1.0.4
mkdocs-material==4.0.2
PyYAML==6.0.3
